--- Src/Components/Atoms/ItemFood/index.js
@@ -1,70 +1,0 @@
-import React from 'react';
-import {StyleSheet, Text, View, Image, TouchableOpacity} from 'react-native';
-import Rating from '../Rating';
-import {Fonts} from '../../../Utils';
-
-const ItemFood = ({onPress, image}) => {
-  return (
-    <TouchableOpacity activeOpacity={0.7} onPress={onPress}>
-      <View style={styles.container}>
-        <View style={styles.wrapLeft}>
-          <View style={styles.wrapImage}>
-            <Image source={image} style={styles.image} />
-          </View>
-          <View style={styles.wrapText}>
-            <Text style={styles.title}>Zuppa Soup</Text>
-            <Text style={styles.price}>Rp. 25.000</Text>
-          </View>
-        </View>
-        <View style={styles.wrapRating}>
-          <Rating />
-        </View>
-      </View>
-    </TouchableOpacity>
-  );
-};
-
-export default ItemFood;
-
-const styles = StyleSheet.create({
-  container: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    alignItems: 'center',
-    paddingVertical: 8,
-  },
-  wrapLeft: {flexDirection: 'row', alignItems: 'center'},
-  wrapImage: {
-    width: 60,
-    height: 60,
-    borderRadius: 8,
-    overflow: 'hidden',
-  },
-  image: {
-    width: 60,
-    height: 60,
-    resizeMode: 'cover',
-    borderRadius: 8,
-  },
-  wrapText: {
-    flexDirection: 'column',
-    marginLeft: 12,
-  },
-  title: {
-    fontSize: 16,
-    fontFamily: Fonts.primary[400],
-    color: 'black',
-    opacity: 0.9,
-  },
-  price: {
-    fontSize: 13,
-    fontFamily: Fonts.primary[400],
-    color: 'black',
-    opacity: 0.6,
-  },
-  wrapRating: {
-    flexDirection: 'row',
-    paddingHorizontal: 6,
-    backgroundColor: 'white',
-  },
-});
