+++ Src/Router/index.js
@@ -38,7 +38,7 @@ const Menu = () => {
 const Router = () => {
   return (
     <Stack.Navigator
-      initialRouteName="Register"
+      initialRouteName="Menu"
       screenOptions={{headerShown: false}}>
       <Stack.Screen component={Menu} name="Menu" />
       <Stack.Screen component={Splash} name="Splash" />
